Time: 24 ms (92.83%), Space: 89.7 MB (30.25%) - LeetHub

+++ 4-median-of-two-sorted-arrays/4-median-of-two-sorted-arrays.cpp
@@ -19,7 +19,7 @@ public:
             l++;
             h--;
         }
-        float median = (nums3[l]+nums3[h])/2;
+        double median = (nums3[l]+nums3[h])/2;
         return median;
